@@ -1,21 +1,19 @@
 name: agora_chat_callkit
-description: A new Flutter package project.
+description: A new Flutter plugin project.
 version: 0.0.1
 homepage:
 
 environment:
   sdk: '>=2.19.1 <3.0.0'
-  flutter: ">=1.17.0"
+  flutter: ">=2.5.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  
+
   agora_chat_sdk: 1.1.0
 
   agora_rtc_engine: 6.1.0
-
-  permission_handler: 10.2.0
 
 dev_dependencies:
   flutter_test:
@@ -27,8 +25,25 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
+  # This section identifies this Flutter project as a plugin project.
+  # The 'pluginClass' specifies the class (in Java, Kotlin, Swift, Objective-C, etc.)
+  # which should be registered in the plugin registry. This is required for
+  # using method channels.
+  # The Android 'package' specifies package in which the registered class is.
+  # This is required for using method channels on Android.
+  # The 'ffiPlugin' specifies that native code should be built and bundled.
+  # This is required for using `dart:ffi`.
+  # All these are used by the tooling to maintain consistency when
+  # adding or updating assets for this project.
+  plugin:
+    platforms:
+      android:
+        package: com.example.agora_chat_callkit
+        pluginClass: AgoraChatCallkitPlugin
+      ios:
+        pluginClass: AgoraChatCallkitPlugin
 
-  # To add assets to your package, add an assets section, like this:
+  # To add assets to your plugin package, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
@@ -39,7 +54,7 @@ flutter:
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
 
-  # To add custom fonts to your package, add a fonts section here,
+  # To add custom fonts to your plugin package, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
